access to pages modified

+++ webapp/category/views.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, redirect,render_template, request
-from flask_login import login_required
+from flask_login import login_required, current_user
 from sqlalchemy import exc
 from webapp.category.models import Category
 from webapp.course.models import Course
@@ -11,17 +11,20 @@ blueprint = Blueprint('category', __name__)
 @login_required 
 def add_a_category():
     if request.method == 'POST':
-        try:
-            name = request.form['category']
-            category_description = 'This is a description'
-            new_category = Category(name, category_description)
-            if new_category.name != '':
-                db.session.add(new_category)
-                db.session.commit()
-        except(exc.IntegrityError):
-            print('the category exists')
+        if current_user.is_admin:
+            try:
+                name = request.form['category']
+                category_description = 'This is a description'
+                new_category = Category(name, category_description)
+                if new_category.name != '':
+                    db.session.add(new_category)
+                    db.session.commit()
+            except(exc.IntegrityError):
+                print('the category exists')
+                return redirect('/categories')     
+            return redirect('/categories') 
+        else:
             return redirect('/categories')     
-        return redirect('/categories') 
     elif request.method == 'GET':
         title = "Categories"
         is_homepage = False
@@ -37,13 +40,16 @@ def add_a_category():
 @blueprint.route('/delete_a_category', methods=['POST'])
 @login_required    
 def delete_a_category():
-    try:
-        category_id = request.form['category_id']
-        Category.query.filter_by(id=category_id).delete() 
-        db.session.commit()
-    except:
-        return redirect('/categories')    
-    return redirect('/categories') 
+    if current_user.is_admin:
+        try:
+            category_id = request.form['category_id']
+            Category.query.filter_by(id=category_id).delete() 
+            db.session.commit()
+        except:
+            return redirect('/categories')    
+        return redirect('/categories') 
+    else:
+        return redirect('/categories')   
 
 @blueprint.route('/<link_path>', methods=['GET', 'POST'])
 @login_required
@@ -51,16 +57,19 @@ def page(link_path):
     category_exists = Category.query.filter_by(name=link_path).first()
     if category_exists:
         if request.method == 'POST':
-            try:
-                name = request.form['course']
-                course_description = 'This is the best course'
-                new_course = Course(name, category_exists.id, course_description)              
-                if new_course.name != '':
-                    db.session.add(new_course)
-                    db.session.commit()
-            except(exc.IntegrityError):
-                print('the course exists')
-            return redirect('/'+link_path) 
+            if current_user.is_admin:
+                try:
+                    name = request.form['course']
+                    course_description = 'This is the best course'
+                    new_course = Course(name, category_exists.id, course_description)              
+                    if new_course.name != '':
+                        db.session.add(new_course)
+                        db.session.commit()
+                except(exc.IntegrityError):
+                    print('the course exists')
+                return redirect('/'+link_path) 
+            else:
+                return redirect('/'+link_path)     
         elif request.method == 'GET':
             title = link_path
             is_homepage = False
