+++ webapp/course/views.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, redirect,render_template, request
-from flask_login import login_required
+from flask_login import login_required, current_user
 from sqlalchemy import exc
 from webapp.category.models import Category
 from webapp.course.models import Course
@@ -16,16 +16,19 @@ def course_page(category_name, course_name):
         course_exists = Course.query.filter(Course.name==course_name, Course.category_id==category_exists.id).first() 
         if course_exists:
             if request.method == 'POST':
-                try:
-                    name = request.form['lesson']
-                    lesson_description = 'This is a lesson description'
-                    new_lesson = Lesson(name, course_exists.id, lesson_description)              
-                    if new_lesson.name != '':
-                        db.session.add(new_lesson)
-                        db.session.commit()
-                except(exc.IntegrityError):
-                    print('the lesson exists')
-                return redirect('/' + category_name + '/' + course_name) 
+                if current_user.is_admin or current_user.is_teacher:
+                    try:
+                        name = request.form['lesson']
+                        lesson_description = 'This is a lesson description'
+                        new_lesson = Lesson(name, course_exists.id, lesson_description)              
+                        if new_lesson.name != '':
+                            db.session.add(new_lesson)
+                            db.session.commit()
+                    except(exc.IntegrityError):
+                        print('the lesson exists')
+                    return redirect('/' + category_name + '/' + course_name) 
+                else:
+                    return redirect('/' + category_name + '/' + course_name)   
             elif request.method == 'GET': 
                 title = course_name
                 is_homepage = False
@@ -45,12 +48,15 @@ def course_page(category_name, course_name):
 @blueprint.route('/delete_a_course', methods=['POST'])
 @login_required    
 def delete_a_course():
-    course_id = request.form['course_id']
-    course_for_delete=Course.query.filter_by(id=course_id).first()
-    category_name=course_for_delete.category.name
-    try:        
-        Course.query.filter_by(id=course_id).delete() 
-        db.session.commit()
-    except:
-        return redirect('/'+category_name)    
-    return redirect('/'+category_name)          
+    if current_user.is_admin:
+        course_id = request.form['course_id']
+        course_for_delete=Course.query.filter_by(id=course_id).first()
+        category_name=course_for_delete.category.name
+        try:        
+            Course.query.filter_by(id=course_id).delete() 
+            db.session.commit()
+        except:
+            return redirect('/'+category_name)    
+        return redirect('/'+category_name)   
+    else:
+        return redirect('/'+category_name)
